encodeHackathon/requirements.txt
requests
azure-ai-textanalytics
flask
configparser
oauth2
tweepy
uuid